EMS-17 Add basic implementation for Credential API for templating code base
Mapper now is a standalone class that contains 2 main mapping function to map object and list.
The reason for this change is that if we want to map a list using ModelMapper, we need to initialize a TypeToken everytime. To solve this, we can create a util with static methods as in the code.

There is also an implementation for global exception handing. The exception should be thrown where it occurs.

Finally, Credential API contains CRUD operations that any simple API will have.