EMS-21 Migrate to Junit5 and implement Credential Integration tests
Maven does not seem to detect JUnit4 tests, so we move to Junit5. Now we can run `mvn test` and maven will take care of the rest.
We also change the code convention for tests:
  - Test classes will end with -Tests
  - Test cases will follow this pattern: should <correct behavior> when <method name> with <fail/success condition>

After unit tests, we should also have integration tests to check if the whole application run correctly. Therefore, this commit also comes with integration tests for Credential module.